1) integrate remotesubnet reconcile with subnet; 2) rename remote.go to remotesubnet.go/remotevtep.go; 3) rename indexByEndpointIPList to indexByEndpointIP
